% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_enrollment.R
\name{add_pct_change}
\alias{add_pct_change}
\title{Add percent change}
\usage{
add_pct_change(df)
}
\arguments{
\item{df}{Tibble containing summarised and filtered data.}
}
\description{
Append column including percent change from previous year.
Only used if selected year is current year and current year only.
}
