% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{admissionsApp-package}
\alias{admissionsApp}
\alias{admissionsApp-package}
\title{admissionsApp: Admissions data application}
\description{
Admissions data exploration app-as-a-package.
}
\author{
\strong{Maintainer}: Jumping Rivers \email{info@jumpingrivers.com}

}
\keyword{internal}
