% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_enrollment.R
\name{calculate_enrollment}
\alias{calculate_enrollment}
\title{Calculate enrollment}
\usage{
calculate_enrollment(
  df,
  input_year,
  input_season,
  input_group,
  input_filter_by,
  input_filter_values,
  pct_change
)
}
\arguments{
\item{df}{Tibble containing daily enrollment data}

\item{input_year}{Year e.g. "2022"}

\item{input_season}{Season e.g. "Fall"}

\item{input_group}{Group e.g. College}

\item{input_filter_by}{Which filters are selected e.g. gender}

\item{input_filter_values}{List of selected values from filters}

\item{pct_change}{Should percent change from previous year be added to data?}
}
\description{
Using user-entered filters and columns for grouping
calculate enrollment headcount across days before term start.
}
