% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_enrollment.R
\name{calculate_headcount}
\alias{calculate_headcount}
\title{Calculate head count}
\usage{
calculate_headcount(df, input_group, input_year, pct_change)
}
\arguments{
\item{df}{Tibble containing filtered enrollment data}

\item{input_group}{Selected input group}

\item{input_year}{Selected input year}

\item{pct_change}{Should percent change from previous year be added to data?}
}
\description{
Sum enrolled students based on selected groupings
to calculate headcount.
}
