% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_line.R
\name{conditional_filter_input}
\alias{conditional_filter_input}
\title{Create conditional filter input}
\usage{
conditional_filter_input(df, col_name, input_label, id, session)
}
\arguments{
\item{df}{Tibble containing retentions data}

\item{col_name}{Which column to create filter for}

\item{input_label}{Label for pickerInput}

\item{id}{ID for namespace}

\item{session}{Shiny session}
}
\description{
Create pickerInput to be used in
conditional panel created by `conditional_filter_panel()`
This function is used in the server (contains `shiny::renderUI`)
}
