% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_over_time_line_chart.R
\name{conditional_filter_panel}
\alias{conditional_filter_panel}
\title{Create conditional filter input}
\usage{
conditional_filter_panel(col_name, panel_name, session)
}
\arguments{
\item{col_name}{Which column to create filter for.}

\item{panel_name}{Name of the panel.}

\item{session}{Shiny session.}
}
\description{
Create conditional panel containing pickerInput
created by `conditional_filter_input()`.
This function is used in the UI (contains `shiny::uiOutput`)
}
