% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_enrollment.R
\name{filter_data}
\alias{filter_data}
\title{Filter data}
\usage{
filter_data(df, input_year, input_season, input_filter_by, input_filter_values)
}
\arguments{
\item{df}{Tibble containing enrollment data}

\item{input_year}{Year e.g. 2022}

\item{input_season}{Season e.g. "Fall"}

\item{input_filter_by}{Which filters are selected e.g. gender}

\item{input_filter_values}{List of selected values from filters}
}
\description{
Filter year, season and any added group filters.
}
