% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_over_time_line_chart.R
\name{generate_line_chart}
\alias{generate_line_chart}
\title{Generate Line Chart}
\usage{
generate_line_chart(
  df,
  x,
  y,
  x_label,
  y_label,
  x_is_continuous = TRUE,
  x_angle = 45,
  x_format = function(x) {
     x
 },
  y_format = function(x) {
     x
 },
  grouping = 0,
  group_labeling = "",
  legend_title = "",
  legend_position = "right",
  lin_reg = FALSE
)
}
\arguments{
\item{df}{A data-frame.}

\item{x, y}{The column of `df` to plot on the x- or y-axis (as a symbol, not a string).}

\item{x_label, y_label}{Labels for the x- or y-axis.}

\item{x_is_continuous}{Is the x-variable a continuous variate?}

\item{x_angle}{Tick-angle for the x-axis.}

\item{x_format, y_format}{Functions for formatting the x- and y- labels.}

\item{grouping}{Column for grouping the plotted data.}

\item{group_labeling}{Label for the groups in the plotted data.}

\item{legend_title}{Title for the legend.}

\item{legend_position}{Position for the legend.}

\item{lin_reg}{Should a linear-regression be overplotted on the graph?}
}
\value{
A ggplotly object.
}
\description{
Generate Line Chart
}
