% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_over_time_line_chart.R
\name{get_enrollment_over_time_df}
\alias{get_enrollment_over_time_df}
\title{Prepare a data-frame to provide data for enrollment line charts}
\usage{
get_enrollment_over_time_df(
  df,
  grouping_selection,
  filter_control,
  filter_values,
  time_col,
  metric_col,
  metric_summarization_function
)
}
\arguments{
\item{df}{A data-frame.}

\item{grouping_selection}{The collection of columns that define the groups to be presented
on the enrollment chart. These are collapsed into a single column 'grouping'.}

\item{filter_control, filter_values}{These define how the rows of `df` are filtered. For each
`df` column-name in `filter_control`, there is a `<column_name>_filter` entry in
`filter_values`. Only rows that match all of the filters will be kept in the summarised output.}

\item{time_col}{Scalar character. Which of the columns defines the enrollment timepoint?}

\item{metric_col}{Scalar character. Which of the columns defines the enrollment metric?}

\item{metric_summarization_function}{How should the entries in `metric_col` be combined?}
}
\description{
This adds a column 'grouping' (defining the lines presented in the plot) and filters rows. Then
summarises enrollment for each time point.
}
