% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_pin_user}
\alias{get_pin_user}
\title{Get user for pinned data}
\usage{
get_pin_user()
}
\description{
Read environment variable which sets
which RStudio Connect account pinned data is read from
}
