% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_pins_board}
\alias{get_pins_board}
\title{Access the pins board that contains data for this app}
\usage{
get_pins_board()
}
\value{
A {pins} board object
}
\description{
If running outside of Connect, this requires the {pins} server/account/key to be defined in the
{golem} config for this app. These values are typically stored in the environment variables
`CONNECT_SERVER`, `CONNECT_ACCOUNT` and `CONNECT_API_KEY`. They are accessed from the fields
`pins_server`, `pins_account`, `pins_key` in the config.
}
\details{
If running on Connect, the server/account/key information is not used.
}
