% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_over_time_line_chart.R
\name{mod_over_time_line_chart_server}
\alias{mod_over_time_line_chart_server}
\title{over_time_line_chart Server Functions}
\usage{
mod_over_time_line_chart_server(
  id,
  df = utShinyMods::entity_time_metric_categories_df,
  time_col = c(Time = "time_column"),
  metric_col = c(Metric = "metric_column"),
  metric_summarization_function = sum,
  grouping_cols = c(`Category 1` = "entity_category_1", `Category 2` =
    "entity_category_2", `Category 3` = "entity_category_3"),
  filter_cols = c(`Category 1` = "entity_category_1", `Category 2` = "entity_category_2",
    `Category 3` = "entity_category_3"),
  module_title = "Title of Module",
  module_sub_title = "Sub Title for module."
)
}
\arguments{
\item{id}{A unique identifier, linking the UI to the Server.}

\item{df}{A data frame.}

\item{time_col}{The time column.}

\item{metric_col}{The metric column.}

\item{metric_summarization_function}{A function for summarising the metric.}

\item{grouping_cols}{The columns to be used for grouping.}

\item{filter_cols}{The columns to be used for filtering.}

\item{module_title}{Title for the chart.}

\item{module_sub_title}{Subtitle for the chart.}
}
\description{
To be copied in the UI
mod_over_time_line_chart_ui("over_time_line_chart_1")
}
\details{
To be copied in the server
mod_over_time_line_chart_server("over_time_line_chart_1")'

Justification for using extra parameters in the Server function, can be found in the following
documentation: https://shiny.rstudio.com/articles/modules.html
Quote from documentation:
"You can define the function so that it takes any number of additional parameters, including ...,
so that whoever uses the module can customize what the module does."
}
