% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_over_time_line_chart.R
\name{mod_over_time_line_chart_ui}
\alias{mod_over_time_line_chart_ui}
\title{over_time_line_chart UI Function}
\usage{
mod_over_time_line_chart_ui(id)
}
\arguments{
\item{id}{A unique identifier, linking the UI to the Server}
}
\description{
A shiny Module.
}
\details{
To be copied in the UI
mod_over_time_line_chart_ui("over_time_line_chart_1")

To be copied in the server
mod_over_time_line_chart_server("over_time_line_chart_1")'
}
