% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{read_enrollment_pin}
\alias{read_enrollment_pin}
\title{Read daily enrollment data from pin}
\usage{
read_enrollment_pin()
}
\value{
data.frame containing the daily enrollment data.
}
\description{
If running outside of Connect, this requires the {pins} server/account/key to be defined in the
{golem} config for this app. The pins key is typically obtained from the environment variable
`RSCONNECT_SERVICE_USER_API_KEY`.
}
