% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_line.R
\name{remaining_group_choices}
\alias{remaining_group_choices}
\title{Create remaining group choices}
\usage{
remaining_group_choices(selected_group = "")
}
\arguments{
\item{selected_group}{Group selected in input widget}
}
\description{
Return all group choices except the selected group.
}
