% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_line.R
\name{this_year}
\alias{this_year}
\title{Return the current year based on the system time}
\usage{
this_year()
}
\value{
The current year as a scalar number.
}
\description{
Return the current year based on the system time
}
