% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_style_utils.R
\name{title_logo}
\alias{title_logo}
\title{Create title logo}
\usage{
title_logo(right_aligned_title, alt = NULL)
}
\arguments{
\item{right_aligned_title}{Text for the title. This is displayed right-aligned on the app.}

\item{alt}{Alt text for the logo.}
}
\value{
Returns UT logo with correct dimensions for app title.
}
\description{
Create title logo
}
