% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_style_utils.R
\name{ut_color_palette}
\alias{ut_color_palette}
\title{Utah Tech Color Palette}
\usage{
ut_color_palette(n)
}
\arguments{
\item{n}{The number of colors to include from the color palette.}
}
\value{
A vector of hex colors, of length n.
}
\description{
A rotating color palette from colors defined by the Utah Tech University branding guide.
}
\examples{
ut_color_palette(2)
ut_color_palette(100)
}
