% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_line.R
\name{conditional_filter_panel}
\alias{conditional_filter_panel}
\title{Create conditional filter input}
\usage{
conditional_filter_panel(col_name, input_label, id)
}
\arguments{
\item{col_name}{Which column to create filter for}

\item{input_label}{Label for pickerInput}

\item{id}{ID for namespace}
}
\description{
Create conditional panel containing pickerInput
created by `conditional_filter_input()`.
This function is used in the UI (contains `shiny::uiOutput`)
}
