% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_daily_enrollment}
\alias{get_daily_enrollment}
\title{Read daily enrollment from pin}
\usage{
get_daily_enrollment()
}
\description{
Read parsed daily enrollment from pin.
Requires env var PIN_USER for path to data file
called daily_enrollment.
}
